fetching PA content before commands that need the latest (#709)
