+++ panther_analysis_tool/command/clone.py
@@ -13,6 +13,7 @@ from panther_analysis_tool.core import (
 
 def run(analysis_id: Optional[str], filter_args: List[str]) -> Tuple[int, str]:
     try:
+        analysis_cache.update_with_latest_panther_analysis(show_progress_bar=True)
         clone(analysis_id, filter_args)
     except (ValueError, analysis_cache.NoCacheException) as err:
         return 1, str(err)
