+++ panther_analysis_tool/command/explore.py
@@ -1,22 +1,30 @@
+from rich.progress import track
+
 from panther_analysis_tool import analysis_utils
 from panther_analysis_tool.core import analysis_cache, versions_file, yaml
 from panther_analysis_tool.gui.explore_gui import ExploreApp
 
 
 def run() -> tuple[int, str]:
-    all_specs = load_all_specs()
+    analysis_cache.update_with_latest_panther_analysis(show_progress_bar=True)
+    all_specs = load_all_specs(show_progress_bar=True)
     app = ExploreApp(all_specs)
     app.run()
     return 0, ""
 
 
-def load_all_specs() -> list[analysis_utils.AnalysisItem]:
+def load_all_specs(show_progress_bar: bool = False) -> list[analysis_utils.AnalysisItem]:
     yaml_loader = yaml.BlockStyleYAML()
     cache = analysis_cache.AnalysisCache()
     versions = versions_file.get_versions().versions
 
     specs: list[analysis_utils.AnalysisItem] = []
-    for _id in cache.list_spec_ids():
+    for _id in track(
+        cache.list_spec_ids(),
+        description="Loading all specs:",
+        disable=not show_progress_bar,
+        transient=True,
+    ):
         spec = cache.get_latest_spec(_id)
         if spec is None:
             continue
