+++ panther_analysis_tool/command/init_project.py
@@ -1,11 +1,13 @@
 import logging
-import os
 import subprocess  # nosec:B404
 from pathlib import Path
 from typing import Tuple
 
+from panther_analysis_tool.core import analysis_cache
+
 
 def run(working_dir: str) -> Tuple[int, str]:
+    analysis_cache.update_with_latest_panther_analysis(show_progress_bar=True)
     setup_git_ignore(Path(working_dir))
     enable_rerere()
     print_ready_message()
@@ -13,42 +15,46 @@ def run(working_dir: str) -> Tuple[int, str]:
 
 
 def setup_git_ignore(working_dir: Path) -> None:
-    if not os.path.exists(working_dir / ".gitignore"):
+    gitignore_file = working_dir / ".gitignore"
+    if not gitignore_file.exists():
         print(".gitignore file created")
-        Path(working_dir / ".gitignore").touch()
+        gitignore_file.touch()
 
-    with open(working_dir / ".gitignore", "r+", encoding="utf-8") as gitignore_file:
-        ignorables = [
-            {
-                "name": "Cache values",
-                "values": [".cache/"],
-            },
-            {
-                "name": "Panther settings",
-                "values": [".panther_settings.*"],
-            },
-            {
-                "name": "Python",
-                "values": ["__pycache__/", "*.pyc"],
-            },
-            {
-                "name": "Panther",
-                "values": ["panther-analysis-*.zip"],
-            },
-            {
-                "name": "IDEs",
-                "values": [".vscode/", ".idea/"],
-            },
-        ]
+    ignorables = [
+        {
+            "name": "Panther settings",
+            "values": [".panther_settings.*"],
+        },
+        {
+            "name": "Python",
+            "values": ["__pycache__/", "*.pyc", ".mypy_cache/", ".pytest_cache/"],
+        },
+        {
+            "name": "Panther",
+            "values": ["panther-analysis-*.zip", ".cache/"],
+        },
+        {
+            "name": "IDEs",
+            "values": [".vscode/", ".idea/"],
+        },
+    ]
 
-        content = gitignore_file.read()
+    content = gitignore_file.read_text()
+
+    # ensure ends with two newlines
+    if content != "" and not content.endswith("\n\n"):
         if not content.endswith("\n"):
-            gitignore_file.write("\n")
-        for ignorable in ignorables:
-            for value in ignorable["values"]:
-                if value not in content:
-                    gitignore_file.write(f"# {ignorable['name']}\n")
-                    gitignore_file.write(f"{value}\n\n")
+            content += "\n"
+        content += "\n"
+
+    for ignorable in ignorables:
+        content += f"# {ignorable['name']}\n"
+        for value in ignorable["values"]:
+            if value not in content:
+                content += f"{value}\n"
+        content += "\n"
+
+    gitignore_file.write_text(content)
 
 
 def enable_rerere() -> None:
@@ -63,9 +69,6 @@ def print_ready_message() -> None:
     print("Project is ready to use!\n")
 
     print("Next, you can start exploring and using Panther out of the box content:")
-    print(
-        "  * Run `pat pull` to pull and merge the latest content from Panther Analysis with your own. Rerun this every time you want to update your content."  # pylint: disable=line-too-long
-    )
     print("  * Run `pat explore` to see the available content.")
     print("  * Run `pat clone <id>` to create a clone of a detection you want to use in your repo.")
     print(
