+++ panther_analysis_tool/constants.py
@@ -95,6 +95,7 @@ CACHE_DIR = pathlib.Path(".cache")
 PANTHER_ANALYSIS_SQLITE_FILE = "panther-analysis.sqlite"
 PANTHER_ANALYSIS_SQLITE_FILE_PATH = CACHE_DIR / PANTHER_ANALYSIS_SQLITE_FILE
 CACHED_VERSIONS_FILE_PATH = CACHE_DIR / ".versions.yml"
+LATEST_CACHED_PANTHER_ANALYSIS_FILE_PATH = CACHE_DIR / "latest-panther-analysis.txt"
 
 DEFAULT_EDITOR = "vi"
