+++ tests/unit/panther_analysis_tool/command/test_pull.py
@@ -1,19 +1,17 @@
 import pathlib
 import shutil
-from unittest import mock
 from unittest.mock import call
 
 from _pytest.monkeypatch import MonkeyPatch
 from pytest_mock import MockerFixture
 
-from panther_analysis_tool import analysis_utils
 from panther_analysis_tool.command import pull
 from panther_analysis_tool.constants import (
     CACHE_DIR,
     CACHED_VERSIONS_FILE_PATH,
     PANTHER_ANALYSIS_SQLITE_FILE_PATH,
 )
-from panther_analysis_tool.core import analysis_cache, versions_file, yaml
+from panther_analysis_tool.core import analysis_cache, yaml
 from panther_analysis_tool.core.git_helpers import CLONED_VERSIONS_FILE_PATH
 
 _FAKE_PY = """
@@ -63,82 +61,6 @@ _FAKE_USER_RULE_2_V1 = yaml.dump(
     }
 )
 
-_FAKE_DATAMODEL_1_V1 = yaml.dump(
-    {
-        "AnalysisType": "datamodel",
-        "Filename": "fake_datamodel_1.py",
-        "DataModelID": "fake.datamodel.1",
-        "Description": "Fake datamodel 1 v1",
-        "Enabled": True,
-    }
-)
-
-_FAKE_LOOKUP_TABLE_1_V1 = yaml.dump(
-    {
-        "AnalysisType": "lookup_table",
-        "LookupName": "fake.lookup_table.1",
-        "Enabled": True,
-        "Description": "Fake lookup table 1 v1",
-    }
-)
-
-_FAKE_GLOBAL_HELPER_1_V1 = yaml.dump(
-    {
-        "AnalysisType": "global",
-        "GlobalID": "fake.global_helper.1",
-        "Enabled": True,
-        "Description": "Fake global helper 1 v1",
-        "Filename": "fake_global_helper_1.py",
-    }
-)
-
-_FAKE_CORRELATION_RULE_1_V1 = yaml.dump(
-    {
-        "AnalysisType": "correlation_rule",
-        "RuleID": "fake.correlation_rule.1",
-        "Enabled": True,
-        "Description": "Fake correlation rule 1 v1",
-    }
-)
-
-_FAKE_POLICY_1_V1 = yaml.dump(
-    {
-        "AnalysisType": "policy",
-        "PolicyID": "fake.policy.1",
-        "Enabled": True,
-        "Description": "Fake policy 1 v1",
-        "Filename": "fake_policy_1.py",
-    }
-)
-
-_FAKE_SCHEDULED_RULE_1_V1 = yaml.dump(
-    {
-        "AnalysisType": "scheduled_rule",
-        "RuleID": "fake.scheduled_rule.1",
-        "Enabled": True,
-        "Description": "Fake scheduled rule 1 v1",
-        "Filename": "fake_scheduled_rule_1.py",
-    }
-)
-
-_FAKE_SAVED_QUERY_1_V1 = yaml.dump(
-    {
-        "AnalysisType": "saved_query",
-        "QueryName": "fake.saved_query.1",
-        "Enabled": True,
-        "Description": "Fake saved query 1 v1",
-    }
-)
-
-_FAKE_SCHEDULED_QUERY_1_V1 = yaml.dump(
-    {
-        "AnalysisType": "scheduled_query",
-        "QueryName": "fake.scheduled_query.1",
-        "Enabled": True,
-        "Description": "Fake scheduled query 1 v1",
-    }
-)
-
 _FAKE_VERSIONS_FILE = yaml.dump(
     {
         "versions": {
@@ -174,106 +96,6 @@ _FAKE_VERSIONS_FILE = yaml.dump(
                     },
                 },
             },
-            "fake.datamodel.1": {
-                "version": 1,
-                "type": "datamodel",
-                "sha256": "fake_sha256_datamodel_1",
-                "history": {
-                    "1": {
-                        "version": 1,
-                        "commit_hash": "fake_commit_hash_datamodel_1",
-                        "yaml_file_path": "fake_datamodel_1.yaml",
-                        "py_file_path": "fake_datamodel_1.py",
-                    },
-                },
-            },
-            "fake.lookup_table.1": {
-                "version": 1,
-                "type": "lookup_table",
-                "sha256": "fake_sha256_lookup_table_1",
-                "history": {
-                    "1": {
-                        "version": 1,
-                        "commit_hash": "fake_commit_hash_lookup_table_1",
-                        "yaml_file_path": "fake_lookup_table_1.yaml",
-                    },
-                },
-            },
-            "fake.global_helper.1": {
-                "version": 1,
-                "type": "global",
-                "sha256": "fake_sha256_global_helper_1",
-                "history": {
-                    "1": {
-                        "version": 1,
-                        "commit_hash": "fake_commit_hash_global_helper_1",
-                        "yaml_file_path": "fake_global_helper_1.yaml",
-                        "py_file_path": "fake_global_helper_1.py",
-                    },
-                },
-            },
-            "fake.correlation_rule.1": {
-                "version": 1,
-                "type": "correlation_rule",
-                "sha256": "fake_sha256_correlation_rule_1",
-                "history": {
-                    "1": {
-                        "version": 1,
-                        "commit_hash": "fake_commit_hash_correlation_rule_1",
-                        "yaml_file_path": "fake_correlation_rule_1.yaml",
-                    },
-                },
-            },
-            "fake.policy.1": {
-                "version": 1,
-                "type": "policy",
-                "sha256": "fake_sha256_policy_1",
-                "history": {
-                    "1": {
-                        "version": 1,
-                        "commit_hash": "fake_commit_hash_policy_1",
-                        "yaml_file_path": "fake_policy_1.yaml",
-                        "py_file_path": "fake_policy_1.py",
-                    },
-                },
-            },
-            "fake.scheduled_rule.1": {
-                "version": 1,
-                "type": "scheduled_rule",
-                "sha256": "fake_sha256_scheduled_rule_1",
-                "history": {
-                    "1": {
-                        "version": 1,
-                        "commit_hash": "fake_commit_hash_scheduled_rule_1",
-                        "yaml_file_path": "fake_scheduled_rule_1.yaml",
-                        "py_file_path": "fake_scheduled_rule_1.py",
-                    },
-                },
-            },
-            "fake.saved_query.1": {
-                "version": 1,
-                "type": "saved_query",
-                "sha256": "fake_sha256_saved_query_1",
-                "history": {
-                    "1": {
-                        "version": 1,
-                        "commit_hash": "fake_commit_hash_saved_query_1",
-                        "yaml_file_path": "fake_saved_query_1.yaml",
-                    },
-                },
-            },
-            "fake.scheduled_query.1": {
-                "version": 1,
-                "type": "scheduled_query",
-                "sha256": "fake_sha256_scheduled_query_1",
-                "history": {
-                    "1": {
-                        "version": 1,
-                        "commit_hash": "fake_commit_hash_scheduled_query_1",
-                        "yaml_file_path": "fake_scheduled_query_1.yaml",
-                    },
-                },
-            },
         },
     }
 )
@@ -292,36 +114,6 @@ def set_up_cache(tmp_path: pathlib.Path, monkeypatch: MonkeyPatch) -> analysis_c
     create_file_with_text(pa_clone_path / "rules" / "fake_rule_1.py", _FAKE_PY)
     create_file_with_text(pa_clone_path / "rules" / "fake_rule_2.yaml", _FAKE_RULE_2_V2)
     create_file_with_text(pa_clone_path / "rules" / "fake_rule_2.py", _FAKE_PY)
-    create_file_with_text(
-        pa_clone_path / "data_models" / "fake_datamodel_1.yaml", _FAKE_DATAMODEL_1_V1
-    )
-    create_file_with_text(pa_clone_path / "data_models" / "fake_datamodel_1.py", _FAKE_PY)
-    create_file_with_text(
-        pa_clone_path / "lookup_tables" / "fake_lookup_table_1.yaml", _FAKE_LOOKUP_TABLE_1_V1
-    )
-    create_file_with_text(
-        pa_clone_path / "global_helpers" / "fake_global_helper_1.yaml", _FAKE_GLOBAL_HELPER_1_V1
-    )
-    create_file_with_text(pa_clone_path / "global_helpers" / "fake_global_helper_1.py", _FAKE_PY)
-    create_file_with_text(
-        pa_clone_path / "correlation_rules" / "fake_correlation_rule_1.yaml",
-        _FAKE_CORRELATION_RULE_1_V1,
-    )
-    create_file_with_text(pa_clone_path / "policies" / "fake_policy_1.yaml", _FAKE_POLICY_1_V1)
-    create_file_with_text(pa_clone_path / "policies" / "fake_policy_1.py", _FAKE_PY)
-    create_file_with_text(
-        pa_clone_path / "scheduled_rules" / "fake_scheduled_rule_1.yaml",
-        _FAKE_SCHEDULED_RULE_1_V1,
-    )
-    create_file_with_text(pa_clone_path / "scheduled_rules" / "fake_scheduled_rule_1.py", _FAKE_PY)
-    create_file_with_text(
-        pa_clone_path / "saved_queries" / "fake_saved_query_1.yaml", _FAKE_SAVED_QUERY_1_V1
-    )
-    create_file_with_text(
-        pa_clone_path / "scheduled_queries" / "fake_scheduled_query_1.yaml",
-        _FAKE_SCHEDULED_QUERY_1_V1,
-    )
-
     create_file_with_text(CACHED_VERSIONS_FILE_PATH, _FAKE_VERSIONS_FILE)
 
     # fake user's analysis items
@@ -338,217 +130,6 @@ def create_file_with_text(path: pathlib.Path, text: str) -> None:
     path.parent.mkdir(parents=True, exist_ok=True)
     path.touch()
     path.write_text(text)
-
-
-def dump_sqlite(cache: analysis_cache.AnalysisCache) -> None:
-    cursor = cache.cursor
-    cursor.execute("SELECT name FROM sqlite_master WHERE type='table';")
-    tables = cursor.fetchall()
-
-    # Dump each table
-    for table in tables:
-        table_name = table[0]
-        print(f"\n=== {table_name} ===")
-        cursor.execute(f"SELECT * FROM {table_name}")
-        rows = cursor.fetchall()
-        for row in rows:
-            print(row)
-
-
-def populate_sqlite_with_test_data(cache: analysis_cache.AnalysisCache) -> None:
-    user_analysis_specs = {
-        spec.analysis_id(): spec for spec in analysis_utils.load_analysis_specs_ex(["."], [], True)
-    }
-    versions = versions_file.get_versions().versions
-
-    for spec in analysis_utils.load_analysis_specs_ex([str(CACHE_DIR)], [], True):
-        pull.populate_sqlite(spec, cache, user_analysis_specs, versions)
-
-
-@mock.patch("panther_analysis_tool.command.pull.git_helpers.get_panther_analysis_file_contents")
-def test_populate_works_with_latest_versions(
-    mock_get_panther_analysis_file_contents: mock.MagicMock,
-    tmp_path: pathlib.Path,
-    monkeypatch: MonkeyPatch,
-) -> None:
-    mock_get_panther_analysis_file_contents.side_effect = [_FAKE_RULE_1_V1, _FAKE_PY]
-    cache = set_up_cache(tmp_path, monkeypatch)
-    populate_sqlite_with_test_data(cache)
-
-    latest_spec = cache.get_latest_spec("fake.rule.1")
-    assert latest_spec is not None
-    assert latest_spec.spec.decode("utf-8") == _FAKE_RULE_1_V1
-    assert latest_spec.version == 1
-    assert latest_spec.id_field == "RuleID"
-    assert latest_spec.id_value == "fake.rule.1"
-
-    py_file = cache.get_file_for_spec(latest_spec.id or -1, latest_spec.version)
-    assert py_file is not None
-    assert py_file.decode("utf-8") == _FAKE_PY
-
-    latest_spec_2 = cache.get_latest_spec("fake.rule.2")
-    assert latest_spec_2 is not None
-    assert latest_spec_2.spec.decode("utf-8") == _FAKE_RULE_2_V2
-    assert latest_spec_2.version == 2
-    assert latest_spec_2.id_field == "RuleID"
-    assert latest_spec_2.id_value == "fake.rule.2"
-
-    py_file = cache.get_file_for_spec(latest_spec.id or -1, latest_spec.version)
-    assert py_file is not None
-    assert py_file.decode("utf-8") == _FAKE_PY
-
-
-@mock.patch("panther_analysis_tool.command.pull.git_helpers.get_panther_analysis_file_contents")
-def test_populate_works_when_user_has_old_version(
-    mock_get_panther_analysis_file_contents: mock.MagicMock,
-    tmp_path: pathlib.Path,
-    monkeypatch: MonkeyPatch,
-) -> None:
-    mock_get_panther_analysis_file_contents.side_effect = [_FAKE_RULE_2_V1, _FAKE_PY + "old"]
-    cache = set_up_cache(tmp_path, monkeypatch)
-    populate_sqlite_with_test_data(cache)
-
-    latest_spec = cache.get_latest_spec("fake.rule.2")
-    assert latest_spec is not None
-    assert latest_spec.spec.decode("utf-8") == _FAKE_RULE_2_V2
-    assert latest_spec.version == 2
-    assert latest_spec.id_field == "RuleID"
-    assert latest_spec.id_value == "fake.rule.2"
-
-    py_file = cache.get_file_for_spec(latest_spec.id or -1, latest_spec.version)
-    assert py_file is not None
-    assert py_file.decode("utf-8") == _FAKE_PY
-
-    old_spec = cache.get_spec_for_version("fake.rule.2", 1)
-    assert old_spec is not None
-    assert old_spec.spec.decode("utf-8") == _FAKE_RULE_2_V1
-    assert old_spec.version == 1
-    assert old_spec.id_field == "RuleID"
-    assert old_spec.id_value == "fake.rule.2"
-
-    py_file = cache.get_file_for_spec(old_spec.id or -1, old_spec.version)
-    assert py_file is not None
-    assert py_file.decode("utf-8") == _FAKE_PY + "old"
-
-
-def test_populate_works_with_datamodel(tmp_path: pathlib.Path, monkeypatch: MonkeyPatch) -> None:
-    cache = set_up_cache(tmp_path, monkeypatch)
-    populate_sqlite_with_test_data(cache)
-
-    latest_spec = cache.get_latest_spec("fake.datamodel.1")
-    assert latest_spec is not None
-    assert latest_spec.spec.decode("utf-8") == _FAKE_DATAMODEL_1_V1
-    assert latest_spec.version == 1
-    assert latest_spec.id_field == "DataModelID"
-    assert latest_spec.id_value == "fake.datamodel.1"
-
-    py_file = cache.get_file_for_spec(latest_spec.id or -1, latest_spec.version)
-    assert py_file is not None
-    assert py_file.decode("utf-8") == _FAKE_PY
-
-
-def test_populate_works_with_lookup_table(tmp_path: pathlib.Path, monkeypatch: MonkeyPatch) -> None:
-    cache = set_up_cache(tmp_path, monkeypatch)
-    populate_sqlite_with_test_data(cache)
-
-    latest_spec = cache.get_latest_spec("fake.lookup_table.1")
-    assert latest_spec is not None
-    assert latest_spec.spec.decode("utf-8") == _FAKE_LOOKUP_TABLE_1_V1
-    assert latest_spec.version == 1
-    assert latest_spec.id_field == "LookupName"
-    assert latest_spec.id_value == "fake.lookup_table.1"
-
-
-def test_populate_works_with_global_helper(
-    tmp_path: pathlib.Path, monkeypatch: MonkeyPatch
-) -> None:
-    cache = set_up_cache(tmp_path, monkeypatch)
-    populate_sqlite_with_test_data(cache)
-
-    latest_spec = cache.get_latest_spec("fake.global_helper.1")
-    assert latest_spec is not None
-    assert latest_spec.spec.decode("utf-8") == _FAKE_GLOBAL_HELPER_1_V1
-    assert latest_spec.version == 1
-    assert latest_spec.id_field == "GlobalID"
-    assert latest_spec.id_value == "fake.global_helper.1"
-
-    py_file = cache.get_file_for_spec(latest_spec.id or -1, latest_spec.version)
-    assert py_file is not None
-    assert py_file.decode("utf-8") == _FAKE_PY
-
-
-def test_populate_works_with_correlation_rule(
-    tmp_path: pathlib.Path, monkeypatch: MonkeyPatch
-) -> None:
-    cache = set_up_cache(tmp_path, monkeypatch)
-    populate_sqlite_with_test_data(cache)
-
-    latest_spec = cache.get_latest_spec("fake.correlation_rule.1")
-    assert latest_spec is not None
-    assert latest_spec.spec.decode("utf-8") == _FAKE_CORRELATION_RULE_1_V1
-    assert latest_spec.version == 1
-    assert latest_spec.id_field == "RuleID"
-    assert latest_spec.id_value == "fake.correlation_rule.1"
-
-
-def test_populate_works_with_policy(tmp_path: pathlib.Path, monkeypatch: MonkeyPatch) -> None:
-    cache = set_up_cache(tmp_path, monkeypatch)
-    populate_sqlite_with_test_data(cache)
-
-    latest_spec = cache.get_latest_spec("fake.policy.1")
-    assert latest_spec is not None
-    assert latest_spec.spec.decode("utf-8") == _FAKE_POLICY_1_V1
-    assert latest_spec.version == 1
-    assert latest_spec.id_field == "PolicyID"
-    assert latest_spec.id_value == "fake.policy.1"
-
-    py_file = cache.get_file_for_spec(latest_spec.id or -1, latest_spec.version)
-    assert py_file is not None
-    assert py_file.decode("utf-8") == _FAKE_PY
-
-
-def test_populate_works_with_scheduled_rule(
-    tmp_path: pathlib.Path, monkeypatch: MonkeyPatch
-) -> None:
-    cache = set_up_cache(tmp_path, monkeypatch)
-    populate_sqlite_with_test_data(cache)
-
-    latest_spec = cache.get_latest_spec("fake.scheduled_rule.1")
-    assert latest_spec is not None
-    assert latest_spec.spec.decode("utf-8") == _FAKE_SCHEDULED_RULE_1_V1
-    assert latest_spec.version == 1
-    assert latest_spec.id_field == "RuleID"
-    assert latest_spec.id_value == "fake.scheduled_rule.1"
-
-    py_file = cache.get_file_for_spec(latest_spec.id or -1, latest_spec.version)
-    assert py_file is not None
-    assert py_file.decode("utf-8") == _FAKE_PY
-
-
-def test_populate_works_with_saved_query(tmp_path: pathlib.Path, monkeypatch: MonkeyPatch) -> None:
-    cache = set_up_cache(tmp_path, monkeypatch)
-    populate_sqlite_with_test_data(cache)
-
-    latest_spec = cache.get_latest_spec("fake.saved_query.1")
-    assert latest_spec is not None
-    assert latest_spec.spec.decode("utf-8") == _FAKE_SAVED_QUERY_1_V1
-    assert latest_spec.version == 1
-    assert latest_spec.id_field == "QueryName"
-    assert latest_spec.id_value == "fake.saved_query.1"
-
-
-def test_populate_works_with_scheduled_query(
-    tmp_path: pathlib.Path, monkeypatch: MonkeyPatch
-) -> None:
-    cache = set_up_cache(tmp_path, monkeypatch)
-    populate_sqlite_with_test_data(cache)
-
-    latest_spec = cache.get_latest_spec("fake.scheduled_query.1")
-    assert latest_spec is not None
-    assert latest_spec.spec.decode("utf-8") == _FAKE_SCHEDULED_QUERY_1_V1
-    assert latest_spec.version == 1
-    assert latest_spec.id_field == "QueryName"
-    assert latest_spec.id_value == "fake.scheduled_query.1"
 
 
 def test_pull_and_merge_works(
@@ -570,37 +151,3 @@ def test_pull_and_merge_works(
             call("  * fake.rule.2"),
         ]
     )
-
-
-def test_populate_skips_packs(tmp_path: pathlib.Path, monkeypatch: MonkeyPatch) -> None:
-    monkeypatch.chdir(tmp_path)
-    PANTHER_ANALYSIS_SQLITE_FILE_PATH.parent.mkdir(parents=True, exist_ok=True)
-    PANTHER_ANALYSIS_SQLITE_FILE_PATH.touch()
-    cache = analysis_cache.AnalysisCache()
-    cache.create_tables()
-    assert cache.list_spec_ids() == []
-
-    spec = analysis_utils.LoadAnalysisSpecsResult(
-        spec_filename="fake.pack.1.yaml",
-        relative_path="fake.pack.1.yaml",
-        analysis_spec={"AnalysisType": "pack", "PackID": "fake.pack.1"},
-        yaml_ctx=yaml.BlockStyleYAML(),
-        error=None,
-        raw_spec_file_content=b"",
-    )
-    versions = {
-        "fake.pack.1": versions_file.AnalysisVersionItem(
-            version=1,
-            sha256="fake.pack.1.sha256",
-            type="pack",
-            history={
-                1: versions_file.AnalysisVersionHistoryItem(
-                    commit_hash="fake.pack.1.commit.hash",
-                    yaml_file_path="fake.pack.1.yaml",
-                    py_file_path="fake.pack.1.py",
-                ),
-            },
-        )
-    }
-    pull.populate_sqlite(spec, cache, {}, versions)
-    assert cache.list_spec_ids() == []
